Renamed Formatter to Serializer #CHANGE: Renamed Formatter classes to more accurate Serializer classes.
