+++ Duality/Duality.csproj
@@ -147,6 +147,7 @@
     <Compile Include="Resources\ShaderInfo\ShaderFieldScope.cs" />
     <Compile Include="Resources\ShaderInfo\ShaderFieldInfo.cs" />
     <Compile Include="Resources\ShaderInfo\ShaderFieldType.cs" />
+    <Compile Include="Serialization\SerializeMethod.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsIColorData.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsStream.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsXml.cs" />
@@ -294,17 +295,17 @@
     <Compile Include="Audio\SoundInstance.cs" />
     <Compile Include="Drawing\FormattedText.cs" />
     <Compile Include="Utility\Log\TextWriterLogOutput.cs" />
-    <Compile Include="Serialization\BinaryFormatter.cs" />
+    <Compile Include="Serialization\BinarySerializer.cs" />
     <Compile Include="Serialization\ISerializeExplicit.cs" />
     <Compile Include="Serialization\IDataWriter.cs" />
     <Compile Include="Serialization\IDataReader.cs" />
     <Compile Include="Serialization\ISerializeSurrogate.cs" />
-    <Compile Include="Serialization\Formatter.cs" />
+    <Compile Include="Serialization\Serializer.cs" />
     <Compile Include="Serialization\IUniqueIdentifyable.cs" />
     <Compile Include="Serialization\ObjectIdManager.cs" />
     <Compile Include="Serialization\SerializeErrorHandler.cs" />
     <Compile Include="Serialization\Surrogates\GuidSurrogate.cs" />
-    <Compile Include="Serialization\XmlFormatter.cs" />
+    <Compile Include="Serialization\XmlSerializer.cs" />
     <Compile Include="Serialization\SerializeType.cs" />
     <Compile Include="Serialization\DataType.cs" />
     <Compile Include="Serialization\Surrogates\BitmapSurrogate.cs" />
