+++ UnitTesting/DualityTests/Serialization/FormatterTest.cs
@@ -13,22 +13,22 @@ using NUnit.Framework;
 
 namespace Duality.Tests.Serialization
 {
-	[TestFixture(FormattingMethod.Xml)]
-	[TestFixture(FormattingMethod.Binary)]
+	[TestFixture(SerializeMethod.Xml)]
+	[TestFixture(SerializeMethod.Binary)]
 	public class FormatterTest
 	{
-		private FormattingMethod format;
+		private SerializeMethod format;
 
-		private FormattingMethod PrimaryFormat
+		private SerializeMethod PrimaryFormat
 		{
 			get { return this.format; }
 		}
-		private IEnumerable<FormattingMethod> OtherFormats
+		private IEnumerable<SerializeMethod> OtherFormats
 		{
-			get { return Enum.GetValues(typeof(FormattingMethod)).Cast<FormattingMethod>().Where(m => m != FormattingMethod.Unknown && m != this.PrimaryFormat); }
+			get { return Enum.GetValues(typeof(SerializeMethod)).Cast<SerializeMethod>().Where(m => m != SerializeMethod.Unknown && m != this.PrimaryFormat); }
 		}
 
-		public FormatterTest(FormattingMethod format)
+		public FormatterTest(SerializeMethod format)
 		{
 			this.format = format;
 		}
@@ -85,7 +85,7 @@ namespace Duality.Tests.Serialization
 			TestObject	dataResult;
 
 			using (MemoryStream stream = new MemoryStream())
-			using (Formatter formatter = Formatter.Create(stream, this.PrimaryFormat))
+			using (Serializer formatter = Serializer.Create(stream, this.PrimaryFormat))
 			{
 				using (BinaryWriter binWriter = new BinaryWriter(stream.NonClosing()))
 				{
@@ -107,7 +107,7 @@ namespace Duality.Tests.Serialization
 			Assert.IsTrue(rawDataB.Equals(rawDataResultB));
 			Assert.IsTrue(data.Equals(dataResult));
 		}
-		[Test] public void ConvertFormat([ValueSource("OtherFormats")] FormattingMethod to)
+		[Test] public void ConvertFormat([ValueSource("OtherFormats")] SerializeMethod to)
 		{
 			Random rnd = new Random();
 			TestObject data = new TestObject(rnd);
@@ -116,27 +116,27 @@ namespace Duality.Tests.Serialization
 			using (MemoryStream stream = new MemoryStream())
 			{
 				// Write old format
-				using (Formatter formatterWrite = Formatter.Create(stream, this.PrimaryFormat))
+				using (Serializer formatterWrite = Serializer.Create(stream, this.PrimaryFormat))
 				{
 					formatterWrite.WriteObject(data);
 				}
 
 				// Read
 				stream.Position = 0;
-				using (Formatter formatterRead = Formatter.Create(stream))
+				using (Serializer formatterRead = Serializer.Create(stream))
 				{
 					formatterRead.ReadObject(out dataResult);
 				}
 
 				// Write new format
-				using (Formatter formatterWrite = Formatter.Create(stream, to))
+				using (Serializer formatterWrite = Serializer.Create(stream, to))
 				{
 					formatterWrite.WriteObject(data);
 				}
 
 				// Read
 				stream.Position = 0;
-				using (Formatter formatterRead = Formatter.Create(stream))
+				using (Serializer formatterRead = Serializer.Create(stream))
 				{
 					formatterRead.ReadObject(out dataResult);
 				}
@@ -168,7 +168,7 @@ namespace Duality.Tests.Serialization
 				// Write
 				for (int i = 0; i < results.Length; i++)
 				{
-					using (Formatter formatterWrite = Formatter.Create(stream, format))
+					using (Serializer formatterWrite = Serializer.Create(stream, format))
 					{
 						formatterWrite.WriteObject(data);
 					}
@@ -180,7 +180,7 @@ namespace Duality.Tests.Serialization
 				// Read
 				for (int i = 0; i < results.Length; i++)
 				{
-					using (Formatter formatterRead = Formatter.Create(stream))
+					using (Serializer formatterRead = Serializer.Create(stream))
 					{
 						results[i] = formatterRead.ReadObject<TestObject>();
 					}
@@ -194,52 +194,52 @@ namespace Duality.Tests.Serialization
 		}
 
 		
-		private string GetReferenceResourceName(string name, FormattingMethod format)
+		private string GetReferenceResourceName(string name, SerializeMethod format)
 		{
 			return string.Format("FormatterTest{0}{1}Data", name, format);
 		}
-		private void CreateReferenceFile<T>(string name, T writeObj, FormattingMethod format)
+		private void CreateReferenceFile<T>(string name, T writeObj, SerializeMethod format)
 		{
 			string filePath = TestHelper.GetEmbeddedResourcePath(GetReferenceResourceName(name, format), ".dat");
 			using (FileStream stream = File.Open(filePath, FileMode.Create))
-			using (Formatter formatter = Formatter.Create(stream, format))
+			using (Serializer formatter = Serializer.Create(stream, format))
 			{
 				formatter.WriteObject(writeObj);
 			}
 		}
 
-		private void TestDataEqual<T>(string name, T writeObj, FormattingMethod format)
+		private void TestDataEqual<T>(string name, T writeObj, SerializeMethod format)
 		{
 			T readObj;
 			byte[] data = (byte[])TestRes.ResourceManager.GetObject(this.GetReferenceResourceName(name, format), System.Globalization.CultureInfo.InvariantCulture);
 			using (MemoryStream stream = new MemoryStream(data))
-			using (Formatter formatter = Formatter.Create(stream, format))
+			using (Serializer formatter = Serializer.Create(stream, format))
 			{
 				formatter.ReadObject(out readObj);
 			}
 			Assert.IsTrue(writeObj.Equals(readObj), "Failed data equality check of Type {0} with Value {1}", typeof(T), writeObj);
 		}
-		private void TestWriteRead<T>(T writeObj, FormattingMethod format)
+		private void TestWriteRead<T>(T writeObj, SerializeMethod format)
 		{
 			T readObj;
 			using (MemoryStream stream = new MemoryStream())
 			{
 				// Write
-				using (Formatter formatterWrite = Formatter.Create(stream, format))
+				using (Serializer formatterWrite = Serializer.Create(stream, format))
 				{
 					formatterWrite.WriteObject(writeObj);
 				}
 
 				// Read
 				stream.Position = 0;
-				using (Formatter formatterRead = Formatter.Create(stream))
+				using (Serializer formatterRead = Serializer.Create(stream))
 				{
 					readObj = formatterRead.ReadObject<T>();
 				}
 			}
 			Assert.IsTrue(writeObj.Equals(readObj), "Failed single WriteRead of Type {0} with Value {1}", typeof(T), writeObj);
 		}
-		private void TestSequential<T>(T writeObjA, T writeObjB, FormattingMethod format)
+		private void TestSequential<T>(T writeObjA, T writeObjB, SerializeMethod format)
 		{
 			T readObjA;
 			T readObjB;
@@ -248,7 +248,7 @@ namespace Duality.Tests.Serialization
 			{
 				long beginPos = stream.Position;
 				// Write
-				using (Formatter formatter = Formatter.Create(stream, format))
+				using (Serializer formatter = Serializer.Create(stream, format))
 				{
 					stream.Position = beginPos;
 					formatter.WriteObject(writeObjA);
@@ -265,7 +265,7 @@ namespace Duality.Tests.Serialization
 
 				// Read
 				stream.Position = beginPos;
-				using (Formatter formatter = Formatter.Create(stream))
+				using (Serializer formatter = Serializer.Create(stream))
 				{
 					readObjA = (T)formatter.ReadObject();
 					readObjB = (T)formatter.ReadObject();
@@ -275,7 +275,7 @@ namespace Duality.Tests.Serialization
 			Assert.IsTrue(writeObjA.Equals(readObjA), "Failed sequential WriteRead of Type {0} with Value {1}", typeof(T), writeObjA);
 			Assert.IsTrue(writeObjB.Equals(readObjB), "Failed sequential WriteRead of Type {0} with Value {1}", typeof(T), writeObjB);
 		}
-		private void TestRandomAccess<T>(T writeObjA, T writeObjB, FormattingMethod format)
+		private void TestRandomAccess<T>(T writeObjA, T writeObjB, SerializeMethod format)
 		{
 			T readObjA;
 			T readObjB;
@@ -285,7 +285,7 @@ namespace Duality.Tests.Serialization
 				long posB = 0;
 				long posA = 0;
 				// Write
-				using (Formatter formatter = Formatter.Create(stream, format))
+				using (Serializer formatter = Serializer.Create(stream, format))
 				{
 					posB = stream.Position;
 					formatter.WriteObject(writeObjB);
@@ -306,7 +306,7 @@ namespace Duality.Tests.Serialization
 				}
 
 				// Read
-				using (Formatter formatter = Formatter.Create(stream, format))
+				using (Serializer formatter = Serializer.Create(stream, format))
 				{
 					stream.Position = posA;
 					readObjA = (T)formatter.ReadObject();
